back_end/myproject/requirements.txt
asgiref==3.7.2
Django==4.2.3
djangorestframework==3.14.0
mysqlclient
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
django-cors-headers
